Q.For Post Request send resource URI once the Student and Q. throw unchecked exceptions.(Both ppt questions)

+++ rest-with-spring/src/main/java/com/restspring/restwithspring/controllers/StudentController.java
@@ -1,6 +1,8 @@
 package com.restspring.restwithspring.controllers;
 
 import com.restspring.restwithspring.entities.Student;
+import com.restspring.restwithspring.exception.StudentAlreadyExistsException;
+import com.restspring.restwithspring.exception.StudentNotFoundException;
 import com.restspring.restwithspring.services.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -17,40 +19,52 @@ public class StudentController {
     @Autowired
     StudentService studentService;
 
+
     @GetMapping("/")
     String getString() {
         return "welcome user";
     }
 
-    //retrieve all students
+
+//    retrieve all students
     @GetMapping("/students")
     List<Student> getAllStudents() {
         return studentService.getAllStudents();
     }
 
-    //retrieve student by id
+
+//    retrieve student by id
     @GetMapping("/students/{id}")
     Student getStudentById(@PathVariable Integer id) {
-        return studentService.getStudentById(id);
+        Student student = studentService.getStudentById(id);
+        if (student == null) {
+            // throwing custom exception
+            throw new StudentNotFoundException("student does not exists");
+        }
+        return student;
     }
 
-    //save student
+
+//    save student
     @PostMapping("/students")
     Student saveStudent(Student student) {
         studentService.saveStudent(student);
         return student;
     }
 
-    //delete student by id
-    @PostMapping("/students/{id}")
+//    delete student by id
+    @DeleteMapping("/students/{id}")
     void deleteStudent(@PathVariable Integer id) {
         Student student = studentService.getStudentById(id);
+        if (student == null) {
+            throw new StudentNotFoundException("student not found");
+        }
         studentService.deleteStudent(student);
     }
 
 
-    //enhanced post request with URI status code
-    @PostMapping(value = "/studentpost",headers = "Accept=application/xml")
+//    enhanced post request with URI status code
+    @PostMapping(value = "/studentpost")
     ResponseEntity<Student> saveStudent1(@RequestBody Student student) {
         studentService.saveStudent(student);
 
@@ -62,7 +76,21 @@ public class StudentController {
 
         return ResponseEntity.created(uri).build();
     }
-    
+
+    @PutMapping("/students/{id}/{name}")
+    public ResponseEntity<Object> updateStudent(@RequestBody Student student, @PathVariable Integer id, @PathVariable String name) {
+
+        Student student5 = studentService.getStudentById(id);
+
+        if (student5 == null) {
+            throw new StudentNotFoundException("student not found");
+        }
+        student5.setName(name);
+
+        studentService.saveStudent(student5);
+
+        return ResponseEntity.noContent().build();
+    }
 
 
 }
